requirements.txt
streamlit
pytesseract
Pillow
opencv-python
spacy
pandas
